+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add fine grained task abortion support.
+
 ### Fixed
 
 - Ensure pool statuses are checked at initialization, `start()` or `destroy()`