+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Ensure pool statuses are checked at initialization, `start()` or `destroy()`
+- Ensure pool statuses are checked at initialization, `start()` or `destroy()`.
+- Sort IWRR worker choice strategy round weights in descending order.
 
 ## [3.0.5] - 2023-10-27