+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.6] - 2023-11-24
+
 ### Fixed
 
 - Ensure pool statuses are checked at initialization, `start()` or `destroy()`.